@@ -9,20 +9,23 @@
     "@emotion/styled": "^11.3.0",
     "@headlessui/react": "^1.3.0",
     "@hookform/resolvers": "^2.6.0",
+    "@knowt/rich-markdown-editor": "github:Knowt/rich-markdown-editor",
+    "@mui/base": "^5.0.0-alpha.77",
     "@mui/icons-material": "^5.0.0",
-    "@mui/lab": "^5.0.0-alpha.48",
+    "@mui/lab": "^5.0.0-alpha.78",
     "@mui/material": "^5.0.0",
     "@mui/styles": "^5.0.0",
-    "@sentry/react": "^6.12.0",
+    "@sentry/react": "^6.19.6",
     "@sentry/tracing": "^6.12.0",
-    "@tailwindcss/typography": "^0.4.1",
+    "@tailwindcss/typography": "^0.5.2",
     "@testing-library/jest-dom": "^5.12.0",
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
-    "algoliasearch": "^4.10.5",
+    "algoliasearch": "^4.13.0",
     "animate.css": "^4.1.1",
     "bad-words": "^3.0.4",
-    "browser-image-compression": "^1.0.15",
+    "browser-image-compression": "^2.0.0",
+    "create-react-app": "^5.0.1",
     "cross-env": "^7.0.3",
     "emailjs-com": "^3.1.0",
     "emoji-picker-react": "^3.4.8",
@@ -31,29 +34,30 @@
     "luxon": "^1.28.0",
     "mixpanel-browser": "^2.41.0",
     "npm": "^7.24.1",
-    "react": "^17.0.2",
+    "react": "^18.0.0",
     "react-animate-on-scroll": "^2.1.5",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.0.0",
     "react-error-boundary": "^3.1.3",
     "react-helmet": "^6.1.0",
-    "react-hook-form": "^7.9.0",
-    "react-hotjar": "^3.0.1",
+    "react-hook-form": "^7.30.0",
+    "react-hotjar": "^5.0.0",
     "react-instantsearch-dom": "^6.12.1",
     "react-intersection-observer": "^8.32.0",
     "react-markdown": "^7.0.1",
+    "react-quill": "^2.0.0-beta.4",
     "react-responsive": "^8.2.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^4.0.3",
+    "react-router-dom": "^6.3.0",
+    "react-router-v6-instrumentation": "^2.0.1",
+    "react-scripts": "^5.0.1",
     "react-select": "^4.3.1",
     "react-select-country-list": "^2.2.3",
     "react-slick": "^0.28.1",
     "react-textarea-autosize": "^8.3.3",
     "react-timeago": "^6.2.1",
     "remark-gfm": "^3.0.0",
-    "rich-markdown-editor": "^11.10.0",
     "slick-carousel": "^1.8.1",
     "styled-components": "^5.3.0",
-    "tailwind-scrollbar-hide": "^1.0.3",
+    "tailwind-scrollbar-hide": "^1.1.7",
     "web-vitals": "^1.1.2",
     "workbox-background-sync": "^5.1.4",
     "workbox-broadcast-update": "^5.1.4",
@@ -70,9 +74,9 @@
     "yup": "^0.32.9"
   },
   "scripts": {
-    "start": "cross-env REACT_APP_VERSION=$(git rev-parse --short HEAD) craco start",
-    "build": "cross-env REACT_APP_VERSION=$(git rev-parse --short HEAD) craco build",
-    "test": "craco test",
+    "start": "cross-env REACT_APP_VERSION=$(git rev-parse --short HEAD) react-scripts start",
+    "build": "cross-env REACT_APP_VERSION=$(git rev-parse --short HEAD) react-scripts build",
+    "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
@@ -94,8 +98,8 @@
     ]
   },
   "devDependencies": {
-    "autoprefixer": "^9.8.6",
-    "postcss": "^7.0.35",
-    "tailwindcss": "npm:@tailwindcss/postcss7-compat@^2.1.4"
+    "autoprefixer": "^10.4.4",
+    "postcss": "^8.4.12",
+    "tailwindcss": "^3.0.24"
   }
 }
