@@ -1,15 +1,16 @@
 import React, { Suspense, lazy, useEffect } from 'react';
 import "./App.css";
-import Home from "./pages/Home/Home";
 import { useAuth } from "./hooks/useAuth";
-import { Router, Switch, Route, Redirect } from "react-router-dom";
-import history from './hooks/history';
+import { Router, Routes, Route, Navigate } from "react-router-dom";
 import ServiceWorkerWrapper from './components/ServiceWorker/ServiceWorkerWrapper';
 import AppLoader from './components/Loader/AppLoader';
 import FallbackPage from './pages/Fallback/FallbackPage';
 import {ErrorBoundary} from 'react-error-boundary';
 import NotFoundPage from './pages/Fallback/NotFoundPage';
 import { logEvent } from './utils/analytics';
+import useRoutingInstrumentation from 'react-router-v6-instrumentation';
+import * as Sentry from '@sentry/react';
+import {Integrations} from '@sentry/tracing';
 
 const LazyLogin = lazy(() => import("./pages/Login/Login"))
 const LazyOnboarding = lazy(() => import("./pages/Onboarding/Onboarding"))
@@ -19,6 +20,20 @@ const LazyAdminApp = lazy(() => import("./container/AdminApp/AdminApp"))
 function App() {
 
   const { user, userDetails } = useAuth();
+  const routingInstrumentation = useRoutingInstrumentation();
+  useEffect(() => {
+    Sentry.init({
+      dsn: "https://0ea940615ee84c89a07bb2e2cc91dbfe@o992434.ingest.sentry.io/5949937",
+      integrations: [new Integrations.BrowserTracing({
+        routingInstrumentation
+      })],
+    
+      // Set tracesSampleRate to 1.0 to capture 100%
+      // of transactions for performance monitoring.
+      // We recommend adjusting this value in production
+      tracesSampleRate: 1.0,
+    });
+  }, [routingInstrumentation]);
 
   useEffect(() => {
     logEvent('app_start', {
@@ -32,25 +47,24 @@ function App() {
 
   return (
     <ErrorBoundary FallbackComponent={FallbackPage} onError={errorHandler}>
-      <Router history={history}>
-        <div className="app bg-gray-100 min-h-screen">
-          <Suspense fallback={<AppLoader/>}>
-            {user !== false? <Switch>
-              <Route exact path={"/login"} component={LazyLogin} />
-              <Route exact path={"/onboarding"} component={LazyOnboarding} /> {/* Onboarding form to get the neccesary details before starting */}
-              <Route path="/admin" component={LazyAdminApp} /> {/* Only render if user is logged in as admin*/}
-              <Route path="/" component={LazyMainApp} /> {/* In the future, we make it so it only renders if user is logged in */}
+      <div className="app bg-gray-100 min-h-screen">
+        <Suspense fallback={<AppLoader/>}>
+          <Routes>
+            {user !== false? <>
+              <Route path="/*" element={<LazyMainApp />}/>
+              <Route path={"/login"} element={<LazyLogin />}/>
+              <Route path="/onboarding" element={<LazyOnboarding />}/>
+              <Route path="/admin/*" element={<LazyAdminApp />}/>
               {/* <Route path="*" component={NotFoundPage}/> */}
-            </Switch>:<Switch>
-              <Route exact path={"/login"} component={LazyLogin} />
-            </Switch>}
-            {user === false && <Route path="/">
-                <Redirect to="/login" />
-              </Route>}
-          </Suspense>
-          <ServiceWorkerWrapper/>
-        </div>
-      </Router>
+            </>: <Route path={"/login"} element={<LazyLogin />}/>}
+            {user === false &&
+            <Route path="/">
+              <Navigate to="/login" />
+            </Route>}
+          </Routes>
+        </Suspense>
+        <ServiceWorkerWrapper/>
+      </div>
     </ErrorBoundary>
   );
 }
