@@ -1,16 +1,14 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
 import { AuthProvider } from "./hooks/useAuth";
 import reportWebVitals from './reportWebVitals';
 import { createTheme, ThemeProvider, StyledEngineProvider, adaptV4Theme } from '@mui/material/styles';
 import * as serviceWorkerRegistration from './serviceWorkerRegistration';
-import * as Sentry from "@sentry/react";
-import { Integrations } from "@sentry/tracing";
-import history from './hooks/history';
+import * as ReactDOM from 'react-dom/client';
+import { BrowserRouter as Router } from 'react-router-dom';
 
-const theme = createTheme(adaptV4Theme({
+const theme = createTheme({
   palette: {
     primary: {
       main: '#084887',
@@ -19,31 +17,23 @@ const theme = createTheme(adaptV4Theme({
       main: '#F58A07',
     },
   },
-}));
-
-Sentry.init({
-  dsn: "https://0ea940615ee84c89a07bb2e2cc91dbfe@o992434.ingest.sentry.io/5949937",
-  integrations: [new Integrations.BrowserTracing({
-    routingInstrumentation: Sentry.reactRouterV5Instrumentation(history),
-  })],
-
-  // Set tracesSampleRate to 1.0 to capture 100%
-  // of transactions for performance monitoring.
-  // We recommend adjusting this value in production
-  tracesSampleRate: 1.0,
 });
 
-ReactDOM.render(
-  <React.StrictMode>
-    <StyledEngineProvider injectFirst>
-      <ThemeProvider theme={theme}>
+
+ReactDOM.createRoot(
+  document.getElementById('root')
+).render(
+<React.StrictMode>
+  <StyledEngineProvider injectFirst>
+    <ThemeProvider theme={theme}>
+      <Router>
         <AuthProvider>
           <App />
         </AuthProvider>
-      </ThemeProvider>
-    </StyledEngineProvider>
-  </React.StrictMode>,
-  document.getElementById('root')
+      </Router>
+    </ThemeProvider>
+  </StyledEngineProvider>
+</React.StrictMode>
 );
 
 // If you want your app to work offline and load faster, you can change
