@@ -4,15 +4,16 @@ import SendIcon from '@mui/icons-material/Send';
 import { db, firebase } from '../../../config/firebase';
 import { useAuth } from '../../../hooks/useAuth';
 import { forwardRef, useState } from 'react';
-import history from '../../../hooks/history';
 import Spinner from '../../Loader/Spinner';
 import { getMessageUserRoom } from '../../../utils/chats';
 import { logEvent } from '../../../utils/analytics';
+import { useNavigate } from 'react-router-dom';
 
 const MentorCard = forwardRef(({ id, name, school, major, bio, highlight1, highlight2, isMtr=true }, ref) => {
     const [message, setMessage] = useState("");
     const { user, userDetails } = useAuth();
     const [sendingMessage, setSendingMessage] = useState(false);
+    const navigate = useNavigate();
 
     const connectWithMentor = async () => {
         /* Send Mentor ID to backend for checking and room creation */
@@ -42,7 +43,7 @@ const MentorCard = forwardRef(({ id, name, school, major, bio, highlight1, highl
                 createdRoom: targetRoomId,
                 sentMessage: message || "Hi 👋"
             })
-            history.push(`/messages/${targetRoomId}`)
+            navigate(`/messages/${targetRoomId}`)
         } catch (e) {
             setSendingMessage(false)
             console.error(e)
@@ -51,13 +52,13 @@ const MentorCard = forwardRef(({ id, name, school, major, bio, highlight1, highl
 
     return (
         <div ref={ref} className="bg-white rounded-lg shadow-lg flex flex-col items-center p-3 mt-16 min-h-80 transition-transform transform hover:scale-105 duration-200 cursor-pointer" style={{flex: "1 0 30%", width: '17.5rem'}}>
-            <ProfilePicture uid={id} alt="mentor" className="w-28 h-28 transform rounded-full -mt-16 border-white border-8 border-solid shadow-lg" style={{aspectRatio: '1'}} onClick={() => history.push(`/profile/${id}`)}/>
+            <ProfilePicture uid={id} alt="mentor" className="w-28 h-28 transform rounded-full -mt-16 border-white border-8 border-solid shadow-lg" style={{aspectRatio: '1'}} onClick={() => navigate(`/profile/${id}`)}/>
             <div className="mt-2 space-y-2 h-full flex flex-col w-full">
-                <div className="space-y-1 text-center transform" onClick={() => history.push(`/profile/${id}`)}>
+                <div className="space-y-1 text-center transform" onClick={() => navigate(`/profile/${id}`)}>
                     <h3 className="font-semibold text-xl">{name}</h3>
                     <p className="text-md text-gray-600">{isMtr?"Mentor":"Student"} @ {school}</p>
                 </div>
-                <div className="space-y-1 text-center flex-1 transform" onClick={() => history.push(`/profile/${id}`)}>
+                <div className="space-y-1 text-center flex-1 transform" onClick={() => navigate(`/profile/${id}`)}>
                     <p className="text-md text-gray-600">{major}</p>
                     <p className="text-sm text-gray-600">{bio.substring(0, 125)}</p>
                     <p>
