@@ -6,12 +6,13 @@ import { db, storage } from '../../config/firebase';
 import { useAuth } from '../../hooks/useAuth';
 import ProfilePicture from '../User/ProfilePicture';
 import Spinner from '../Loader/Spinner';
-import history from '../../hooks/history';
 import { Button, Collapse } from '@mui/material';
 import { TransitionGroup } from "react-transition-group";
+import { useLocation, useNavigate } from 'react-router-dom';
 
 function Sidebar({screenSize}) {
-
+    const location = useLocation();
+    const navigate = useNavigate();
     const { user, userDetails } = useAuth();
 
     const [rooms, setRooms] = useState([]);
@@ -48,8 +49,8 @@ function Sidebar({screenSize}) {
                     }
                 }
             })).then(res => {
-                if(res.length > 0 && history.location.pathname === '/messages' && screenSize >= 1) {
-                    history.push(`/messages/${res[0].id}`);
+                if(res.length > 0 && location.pathname === '/messages' && screenSize >= 1) {
+                    navigate(`/messages/${res[0].id}`);
                 }
                 setLoading(false)
                 setRooms(res)
@@ -100,7 +101,7 @@ function Sidebar({screenSize}) {
                 <div className="h-full grid place-items-center">
                     <div className="space-y-3 grid place-items-center">
                         <h3 className="text-gray-500">No Contacts yet</h3>
-                        <Button variant="contained" color="primary" onClick={() => history.push('/mentors')}>Find a mentor</Button>
+                        <Button variant="contained" color="primary" onClick={() => navigate('/mentors')}>Find a mentor</Button>
                     </div>
                 </div>}
                 {loading&& <div className="grid place-items-center w-full h-14">
