@@ -1,17 +1,15 @@
 import React, { useState, useEffect, forwardRef } from 'react';
 import "./SidebarChat.css";
 import { Avatar } from "@mui/material";
-import { Link, useRouteMatch } from 'react-router-dom';
+import { Link, useLocation, useMatch, useNavigate } from 'react-router-dom';
 import { db } from '../../config/firebase';
 import { useAuth } from '../../hooks/useAuth';
-import history from '../../hooks/history';
 import FiberManualRecordIcon from '@mui/icons-material/FiberManualRecord';
 
 const SidebarChat = forwardRef(({ id, roomName, isMentor, roomPic, read }, ref) =>{
     const { user } = useAuth();
     const [messages, setMessages] = useState('');
-    let { url } = useRouteMatch();
-
+    const location = useLocation();
     //Fetches the latest message for display
     useEffect(() => {
         if (id) {
@@ -22,8 +20,8 @@ const SidebarChat = forwardRef(({ id, roomName, isMentor, roomPic, read }, ref) 
     }, [id])
 
     return (
-        <Link to={`${url}/${id}`}>
-            <div ref={ref} className={`flex items-center p-3 pr-5 ${(`${url}/${id}` === history.location.pathname)?'bg-gray-100':''} transition-colors hover:bg-gray-100`}>
+        <Link to={`${id}`}>
+            <div ref={ref} className={`flex items-center p-3 pr-5 ${(`/messages/${id}` === location.pathname)?'bg-gray-100':''} transition-colors hover:bg-gray-100`}>
                 <Avatar src={roomPic}/> {/* Add src={*room_pic*} in the Avatar tag. Room pic defaults to the other user's prof pic if there are
                 only 2 users, and  the group pic if it is a group chat. <-- Implement this to come from Sidebar and be passed down as a prop */}
                 <div className="sidebarChat__info">
