@@ -7,7 +7,7 @@ function FallbackPage({ error }) {
                 <img className="object-contain w-24 p-4" src="/rocket-logo@3x.png" />
                 <h2>There was an Error</h2>
                 <p>Please Refresh The Page</p>
-                {error}
+                {error.toString()}
                 <div className="animate-pulse h-2 rounded-md w-52 bg-gray-400"></div>
             </div>
         </div>
